split completion text to sequence_len
